// engine/bomberland-library/src/Types/Connection.types.ts
export enum PacketType {
    Bomb = "bomb",
    Detonate = "detonate",
    EndGameState = "endgame_state",
    EvaluateNextState = "evaluate_next_state",
    GameState = "game_state",
    Info = "info",
    Move = "move",
    NextGameState = "next_game_state",
    RequestGameReset = "request_game_reset",
    RequestTick = "request_tick",
    Tick = "tick",
}